requirements.txt
opencv-python
pillow
numpy
requests
python-dotenv
easyocr
pytesseract
matplotlib
regex 